Fix terraform recreating route table rules on every apply
This fixes terraform always destroying and recreating the `azurerm_route`s on every single apply, even if they haven't changed.

Azure automatically creates a route table for our kubernetes cluster which we then add routes to using terraform. To get that route table id we have a `data.azurerm_subnet.cluster_nodes` which has a `route_table_id` attribute. However if we read from that data source before the cluster (and its route table) has been created `route_table_id` is `null`, so we used to mark the data source as required the cluster being created first using as explicit `depends_on` field.

However it turns out that having the explicit `depends_on` means terraform will [never read from the data source during the planning step](https://github.com/hashicorp/terraform/issues/22005#issuecomment-509352901). This means it doesn't have enough information to know whether the current routes are up-to-date or not, so it always puts in the plan to destroy and recreate them.

So to create an implicit dependency on the cluster instead I've parsed the subnet name out the cluster's `vnet_subnet_id` attribute, so that now terraform will hopefully still wait for the cluster to be created before creating the routes, and it won't needlessly recreate the routes all the time.