Add frontend UI for territory filtering
The method we're using to filter territories is similar to how we're
currently filtering documents.

We've added the filter checkboxes to the left hand side of the page,
and when each filter is selected it is appended to a query string in the
page's URL.

The implementation of this territory filtering has been mostly
generalising the current implementation of filtering documents. We're
then able to re-use most of the existing functionality by passing in the
existing `DocType` or new `TerritoryType`.

Closes #1123.