Add territory to doc-index-updater and search
Now that `territory` has been added as a field in Sentinel, we want to
process this value when passed into the `doc-index-updater`.

This commit updates both the `doc-index-updater` and `search` projects.

For `doc-index-updater`, we've updated the `Document` struct with
`territory`, and all of its dependent code references and tests. The
`audit_logger`, `create_manager` and `delete_manager` should now all be
aware of this new field.

For `search` we've added `territory` to the default search index. We've
also added some better error handling for the methods that communicate
with Azure's REST interface, most notably using PUT requests rather than
POST requests to create and update indexes.