Derive territory on import
Azure's search doesn't allow us to prefix search the `pl_number` field.
Therefore in order to search for a territory through the `pl_number`
prefix, we derive the `territory` from the `pl_number` during import,
but only if a `territory` is not already given.

The basic rules are:

If territory is present in document:
  return territory
Else:
  If `pl_number` prefix for known territory type:
    return matching territory type
  Else:
    return None