Fix GB & NI territory filtering
For GB and NI territories we were filtering so that it would only return
GB and NI territory results respectively. This has now been changed so
that for GB we were return GB AND UK results, and for NI we return NI
AND UK results.

This commit updates these queries in both the graphQL API, as well a
the direct Azure search.